Bring current state of the private repository
The original repository has a history with content that might be
incompatible with the Unlicense license, so I moved the current stable
state while keeping that out.

The dependency constraints were kept the same, but updated to the latest
compatible version.

Also, a LICENSE and basic README are added with this commit.