Replace `dirs-next` with `directories`
The crate `dirs-next` is abandoned, the last update in crates.io was 2
years ago and the latest commit on the `xdg-rs/dirs` repository was on
2021, its README points to the `dirs-dev` organization where the `dirs`
crate resides, which has a very similar API to `dirs-next`.

And while that should be it, its README recommends the use of the
`directories` crate for application related directories, which besides
making me see that having proper directories for different operating
systems has more nuances that I thought, it also highlighted what I
found myself doing with `dirs-next`: getting the path and suffixing it
with a project directory to get a root to join its contents to, and
`directories` takes care of that.

The `lazy_static` dependency was added, since it is required to compute
all the paths to start using them and I don't necessarily see myself
needing to retrieve them more than once during a single execution of the
application, why would I need the most up-to-date paths for this anyway?

As a bonus, once initialized the config and data directories are not
`Option`al results, that seems like a plus in my book.