Make error messages more user-friendly
This is pretty loaded one.

It started with wanting an anyhow/eyre like message, with the error and
the chain of sources listed just after. But then I took advantage of
using my own defined error type, `TrackerError`, to give suggestions
and/or possible fixes.

For the sources, I had the problem of not knowing what to do to return
the boxed error coming from the configuration, then I noticed that they
could only be two types. Why not use an enum to keep the types? It even
ensures that a path is given when the error originated from parsing a
file!

Then the `TrackerFormat` variant was added to construct a message for
those cases. Came from legacy? Migrate! Unexpected serialization error?
Report it!

While making the `UnexpectedResponse` I noticed that it would be handy
to have the story ID that caused it. Wait, I can! And that is why the
`id` field was added to it.

The plan was to also get `Request` its own messages, but I definitely
don't know how to handle them. It stays to make it easier to do it
later.

I also wanted the messages to be consistent, resulting in almost all of
the `context` messages to be modified.

Maybe there is other stuff that I forget to explain the why, but this
has dragged on for longer than I wanted it to. I want to just move on
for now to the other things that need to be touched upon.

What a mess.