Ensure `download` makes use of the save on error
This completes the last commit, specially while using the `download`
command which only inserted the changes after all downloads were done.
This commit mainly fixes this.

While on this, the `download_stories` macro was removed seeing as it was
difficult to adapt it for it to be used in the `download` subcommand, so
it ended up discarded and just repeating what needed to be repeated in
the `track` command. Bringing the `download_delay` macro for the segment
that is exactly the same.

With that said, this needs testing to ensure that it works as intended.