Make the use of variables in `exec` simpler
Since the split was being done after the variable substitution, any
space present in them would separate the variable into multiple
arguments making it required to enclose in quotes to ensure it stays as
one argument after the split.

The solution applied was to invert the order, first split into arguments
and then substitute the variables. This ends up reducing the amount of
quotes used in `exec`.

Also, this moves this process into its own function to allow for easier
testing of the results.